requirements.txt
streamlit>=1.25.0
numpy>=1.21.0
scikit-learn>=1.0.0
Pillow>=9.0.0
scikit-image>=0.19.0
joblib>=1.1.0
